+++ packages/gcf-utils/test/gcf-bootstrapper.ts
@@ -41,6 +41,15 @@ function nockListInstallationRepos() {
   );
 }
 
+function nockListInstallations(fixture = 'app_installations.json') {
+  return (
+    nock('https://api.github.com/')
+      .get('/app/installations')
+      // eslint-disable-next-line @typescript-eslint/no-var-requires
+      .reply(200, require(`../../test/fixtures/${fixture}`))
+  );
+}
+
 describe('GCFBootstrapper', () => {
   describe('gcf', () => {
     let handler: (
@@ -58,7 +67,10 @@ describe('GCFBootstrapper', () => {
 
     let req: express.Request;
 
-    const spy: sinon.SinonStub = sinon.stub();
+    const issueSpy: sinon.SinonStub = sinon.stub();
+    const repositoryCronSpy: sinon.SinonStub = sinon.stub();
+    const installationCronSpy: sinon.SinonStub = sinon.stub();
+    const globalCronSpy: sinon.SinonStub = sinon.stub();
     let configStub: sinon.SinonStub<[boolean?], Promise<Options>>;
 
     let bootstrapper: GCFBootstrapper;
@@ -91,9 +103,13 @@ describe('GCFBootstrapper', () => {
         .stub(bootstrapper, 'getAuthenticatedOctokit')
         .resolves(new Octokit());
       handler = bootstrapper.gcf(async app => {
-        app.on('issues', spy);
+        app.on('issues', issueSpy);
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        app.on('schedule.repository' as any, spy);
+        app.on('schedule.repository' as any, repositoryCronSpy);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        app.on('schedule.installation' as any, installationCronSpy);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        app.on('schedule.global' as any, globalCronSpy);
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         app.on('err' as any, sinon.stub().throws());
       }, wrapOpts);
@@ -102,7 +118,10 @@ describe('GCFBootstrapper', () => {
     afterEach(() => {
       sendStub.reset();
       sendStatusStub.reset();
-      spy.reset();
+      issueSpy.reset();
+      repositoryCronSpy.reset();
+      installationCronSpy.reset();
+      globalCronSpy.reset();
       configStub.reset();
       enqueueTask.reset();
     });
@@ -123,7 +142,7 @@ describe('GCFBootstrapper', () => {
       sinon.assert.calledOnce(configStub);
       sinon.assert.notCalled(sendStatusStub);
       sinon.assert.calledOnce(sendStub);
-      sinon.assert.calledOnce(spy);
+      sinon.assert.calledOnce(issueSpy);
     });
 
     it('does not schedule task if background option is "false"', async () => {
@@ -143,7 +162,7 @@ describe('GCFBootstrapper', () => {
       sinon.assert.calledOnce(configStub);
       sinon.assert.notCalled(sendStatusStub);
       sinon.assert.calledOnce(sendStub);
-      sinon.assert.calledOnce(spy);
+      sinon.assert.calledOnce(issueSpy);
     });
 
     it('does nothing if there are missing headers', async () => {
@@ -156,15 +175,17 @@ describe('GCFBootstrapper', () => {
       await handler(req, response);
 
       sinon.assert.calledOnce(configStub);
-      sinon.assert.notCalled(spy);
-      sinon.assert.notCalled(sendStub);
+      sinon.assert.notCalled(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
       sinon.assert.calledWith(sendStatusStub, 400);
     });
 
     it('returns 500 on errors', async () => {
       await mockBootstrapper();
       req.body = {
-        installtion: {id: 1},
+        installation: {id: 1},
       };
       req.headers = {};
       req.headers['x-github-event'] = 'err';
@@ -174,7 +195,10 @@ describe('GCFBootstrapper', () => {
       await handler(req, response);
 
       sinon.assert.calledOnce(configStub);
-      sinon.assert.notCalled(spy);
+      sinon.assert.notCalled(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
       sinon.assert.notCalled(sendStatusStub);
       sinon.assert.called(sendStub);
     });
@@ -193,6 +217,33 @@ describe('GCFBootstrapper', () => {
       await handler(req, response);
 
       sinon.assert.calledOnce(enqueueTask);
+      sinon.assert.notCalled(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
+    });
+
+    it('ensures that task is enqueued when called by scheduler for a bot opt out from background execution', async () => {
+      await mockBootstrapper({
+        background: false,
+        logging: true,
+      });
+      req.body = {
+        installation: {id: 1},
+        repo: 'firstRepo',
+      };
+      req.headers = {};
+      req.headers['x-github-event'] = 'schedule.repository';
+      req.headers['x-github-delivery'] = '123';
+      req.headers['x-cloudtasks-taskname'] = '';
+
+      await handler(req, response);
+
+      sinon.assert.calledOnce(enqueueTask);
+      sinon.assert.notCalled(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
     });
 
     it('stores task payload in Cloud Storage if WEBHOOK_TMP set', async () => {
@@ -233,6 +284,10 @@ describe('GCFBootstrapper', () => {
       // We should be attempting to write req.body to Cloud Storage:
       assert.strictEqual(uploaded?.installation?.id, 1);
       sinon.assert.calledOnce(enqueueTask);
+      sinon.assert.notCalled(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
       upload.done();
     });
 
@@ -267,30 +322,274 @@ describe('GCFBootstrapper', () => {
       sinon.assert.calledOnce(configStub);
       sinon.assert.notCalled(sendStatusStub);
       sinon.assert.calledOnce(sendStub);
-      sinon.assert.calledOnce(spy);
+      sinon.assert.calledOnce(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
       downloaded.done();
     });
 
-    it('ensures that task is enqueued when called by scheduler for many repos', async () => {
+    it('does not retry the task, if tmpUrl in payload cannot be found (expired)', async () => {
       await mockBootstrapper();
+      process.env.WEBHOOK_TMP = '/tmp/foo';
       req.body = {
+        tmpUrl: '/bucket/foo',
         installation: {id: 1},
       };
       req.headers = {};
-      req.headers['x-github-event'] = 'schedule.repository';
+      req.headers['x-github-event'] = 'issues';
       req.headers['x-github-delivery'] = '123';
-      req.headers['x-cloudtasks-taskname'] = '';
-      nockListInstallationRepos();
+      // populated once this job has been executed by cloud tasks:
+      req.headers['x-cloudtasks-taskname'] = 'my-task';
+      // Fake download from Cloud Storage, again with the goal of ensuring
+      // we're using the streams API appropriately:
+      const downloaded = nock('https://storage.googleapis.com')
+        .get('/storage/v1/b/tmp/foo/o/%2Fbucket%2Ffoo?alt=media')
+        .reply(404);
 
       await handler(req, response);
 
-      sinon.assert.calledTwice(enqueueTask);
+      delete process.env.WEBHOOK_TMP;
+      sinon.assert.calledOnce(configStub);
+      sinon.assert.notCalled(sendStatusStub);
+      sinon.assert.calledOnceWithMatch(sendStub, {statusCode: 200});
+      sinon.assert.notCalled(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
+      downloaded.done();
+    });
+
+    it('retries the task, if tmpUrl in payload cannot be fetched for other reason', async () => {
+      await mockBootstrapper();
+      process.env.WEBHOOK_TMP = '/tmp/foo';
+      req.body = {
+        tmpUrl: '/bucket/foo',
+        installation: {id: 1},
+      };
+      req.headers = {};
+      req.headers['x-github-event'] = 'issues';
+      req.headers['x-github-delivery'] = '123';
+      // populated once this job has been executed by cloud tasks:
+      req.headers['x-cloudtasks-taskname'] = 'my-task';
+      // Fake download from Cloud Storage, again with the goal of ensuring
+      // we're using the streams API appropriately:
+      const downloaded = nock('https://storage.googleapis.com')
+        .get('/storage/v1/b/tmp/foo/o/%2Fbucket%2Ffoo?alt=media')
+        .reply(500);
+
+      await handler(req, response);
+
+      delete process.env.WEBHOOK_TMP;
+      sinon.assert.calledOnce(configStub);
+      sinon.assert.notCalled(sendStatusStub);
+      sinon.assert.calledOnceWithMatch(sendStub, {statusCode: 500});
+      sinon.assert.notCalled(issueSpy);
+      sinon.assert.notCalled(repositoryCronSpy);
+      sinon.assert.notCalled(installationCronSpy);
+      sinon.assert.notCalled(globalCronSpy);
+      downloaded.done();
+    });
+
+    describe('per-repository cron', () => {
+      it('ensures that task is enqueued when called by scheduler for many repos', async () => {
+        await mockBootstrapper();
+        req.body = {
+          installation: {id: 1},
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.repository';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = '';
+        const listInstallationRepoRequests = nockListInstallationRepos();
+
+        await handler(req, response);
+
+        sinon.assert.calledTwice(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+        listInstallationRepoRequests.done();
+      });
+
+      it('ensures that task is enqueued when called by scheduler for many installations', async () => {
+        await mockBootstrapper();
+        req.body = {};
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.repository';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = '';
+        const listInstallationRequests = nockListInstallations();
+        const listInstallationRepoRequests = nockListInstallationRepos();
+
+        await handler(req, response);
+
+        sinon.assert.calledTwice(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+        listInstallationRequests.done();
+        listInstallationRepoRequests.done();
+      });
+
+      it('handles the schedule.repository task', async () => {
+        await mockBootstrapper();
+        req.body = {
+          repo: 'test-owner/test-repo',
+          installation: {id: 1},
+          cron_type: 'repository',
+          cron_org: 'some-cron-org',
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.repository';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = 'test-function';
+
+        await handler(req, response);
+
+        sinon.assert.notCalled(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.calledOnce(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+      });
+    });
+
+    describe('per-installation cron', () => {
+      it('ensures that task is enqueued when called by scheduler', async () => {
+        await mockBootstrapper();
+        req.body = {
+          cron_type: 'installation',
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.repository';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = '';
+        const listInstallationRequests = nockListInstallations(
+          'app_installations.json'
+        );
+
+        await handler(req, response);
+
+        sinon.assert.calledOnce(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+        listInstallationRequests.done();
+      });
+
+      it('ensures that task is enqueued when called by scheduler for many installations', async () => {
+        await mockBootstrapper();
+        req.body = {
+          cron_type: 'installation',
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.repository';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = '';
+        const listInstallationRequests = nockListInstallations(
+          'app_installations_multiple.json'
+        );
+
+        await handler(req, response);
+
+        sinon.assert.calledTwice(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+        listInstallationRequests.done();
+      });
+
+      it('ensures that task is enqueued when called by scheduler with an installation id', async () => {
+        await mockBootstrapper();
+        req.body = {
+          cron_type: 'installation',
+          installation: {
+            id: 1,
+          },
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.repository';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = '';
+
+        await handler(req, response);
+
+        sinon.assert.calledOnce(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+      });
+
+      it('handles the schedule.installation task', async () => {
+        await mockBootstrapper();
+        req.body = {
+          installation: {id: 1},
+          cron_type: 'installation',
+          cron_org: 'some-cron-org',
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.installation';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = 'test-function';
+
+        await handler(req, response);
+
+        sinon.assert.notCalled(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.calledOnce(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+      });
+    });
+
+    describe('global cron', () => {
+      it('enqueues a single task for global scheduled task', async () => {
+        await mockBootstrapper();
+        req.body = {
+          cron_type: 'global',
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.repository';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = '';
+        await handler(req, response);
+        sinon.assert.calledOnce(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.notCalled(globalCronSpy);
+      });
+
+      it('handles the schedule.global task', async () => {
+        await mockBootstrapper();
+        req.body = {
+          cron_type: 'global',
+        };
+        req.headers = {};
+        req.headers['x-github-event'] = 'schedule.global';
+        req.headers['x-github-delivery'] = '123';
+        req.headers['x-cloudtasks-taskname'] = 'test-function';
+
+        await handler(req, response);
+
+        sinon.assert.notCalled(enqueueTask);
+        sinon.assert.notCalled(issueSpy);
+        sinon.assert.notCalled(repositoryCronSpy);
+        sinon.assert.notCalled(installationCronSpy);
+        sinon.assert.calledOnce(globalCronSpy);
+      });
     });
 
     it('ensures that task is enqueued when called by Github', async () => {
       await mockBootstrapper();
       req.body = {
-        installtion: {id: 1},
+        installation: {id: 1},
       };
       req.headers = {};
       req.headers['x-github-event'] = 'another.name';